salt/minion: allow to customize unit via pillars
to override the unit options of the salt-minion service unit, re-use the
`override.conf` file introduced for the cachecleaner functionality.
Move the management of the `override.conf` out from the SLS
salt.minion.cachecleaner and place it in salt.minion instead, where it
is rendered based on data passed down through
`saltdata.minion.service.unitoptions`, instead of writing hard-coded
values as before.
The options required for the cachecleaner functionality are now
processed in `map.jinja`, where they're added to `unitoptions` as
needed.