salt/common: fix pkgrepo handling for undefined distributions
Up to now, the `pkgrepo` state expected to find an SLS based on the
distribution's name in `salt/common/pkgrepo`.

As some distributions might provide their own Salt packages, it might be
not always needed to configure a separate package repository.

Check first whether a corresponding SLS exists, before trying to include
it.